LayoutTests/fast/block/over-constrained-auto-margin-expected.txt

PASS getComputedStyle(target1).marginRight is '-116px'
PASS getComputedStyle(target2).marginLeft is '-116px'
